requirements.txt
# Core dependencies (cross-platform)
opencv-python>=4.5.0
mediapipe>=0.10.0
pyautogui>=0.9.53
pyyaml>=6.0
pillow>=9.0

# Optional: platform-specific controls
# pycaw  # Windows volume control
# screen-brightness-control  # Brightness control
